+++ lib/ts/recipe/thirdpartypasswordless/recipeImplementation/passwordlessImplementation.ts
@@ -1,17 +1,17 @@
 import { RecipeInterface as WebJSPasswordlessRecipeInterface } from "supertokens-web-js/recipe/passwordless";
-import { TPPWlessRecipeInterface } from "../types";
+import { RecipeInterface as TPPWlessRecipeInterface } from "supertokens-web-js/recipe/thirdpartypasswordless";
 
 export default function getRecipeImplementation(
     originalImplementation: TPPWlessRecipeInterface
 ): WebJSPasswordlessRecipeInterface {
     return {
         clearLoginAttemptInfo: originalImplementation.clearPasswordlessLoginAttemptInfo.bind(originalImplementation),
-        consumeCode: originalImplementation.consumeCode.bind(originalImplementation),
-        createCode: originalImplementation.createCode.bind(originalImplementation),
+        consumeCode: originalImplementation.consumePasswordlessCode.bind(originalImplementation),
+        createCode: originalImplementation.createPasswordlessCode.bind(originalImplementation),
         doesEmailExist: originalImplementation.doesPasswordlessUserEmailExist.bind(originalImplementation),
         doesPhoneNumberExist: originalImplementation.doesPasswordlessUserPhoneNumberExist.bind(originalImplementation),
         getLoginAttemptInfo: originalImplementation.getPasswordlessLoginAttemptInfo.bind(originalImplementation),
-        resendCode: originalImplementation.resendCode.bind(originalImplementation),
+        resendCode: originalImplementation.resendPasswordlessCode.bind(originalImplementation),
         setLoginAttemptInfo: originalImplementation.setPasswordlessLoginAttemptInfo.bind(originalImplementation),
         getLinkCodeFromURL: originalImplementation.getPasswordlessLinkCodeFromURL.bind(originalImplementation),
         getPreAuthSessionIdFromURL: