+++ lib/ts/recipe/thirdpartypasswordless/types.ts
@@ -37,7 +37,6 @@ import {
     PreAPIHookContext as ThirdPartyPreAPIHookContext,
 } from "../thirdparty";
 import {
-    StateObject,
     ThirdPartySignInAndUpState,
     ThirdPartySignInUpActions,
     ThirdPartySignInUpChildProps,
@@ -63,9 +62,7 @@ import { Header } from "./components/themes/signInUp/header";
 import { CountryCode } from "libphonenumber-js";
 import { Dispatch } from "react";
 import { SignInUpScreens } from "../passwordless/components/themes/signInUp";
-import { PasswordlessUser, RecipeFunctionOptions } from "supertokens-web-js/recipe/passwordless";
-import { PasswordlessFlowType } from "supertokens-web-js/lib/build/recipe/passwordless/types";
-import { UserType } from "supertokens-web-js/lib/build/recipe/authRecipeWithEmailVerification/types";
+import { RecipeInterface } from "supertokens-web-js/recipe/thirdpartypasswordless";
 
 // This defines a new type by renaming a single K property to L and as optional
 // If we upgrade to typescript 4.1> this could just keep the optionality of the original (with the as keyword)
@@ -123,9 +120,9 @@ export type UserInput = (
 ) & {
     override?: {
         functions?: (
-            originalImplementation: TPPWlessRecipeInterface,
-            builder?: OverrideableBuilder<TPPWlessRecipeInterface>
-        ) => TPPWlessRecipeInterface;
+            originalImplementation: RecipeInterface,
+            builder?: OverrideableBuilder<RecipeInterface>
+        ) => RecipeInterface;
         components?: ComponentOverrideMap;
     } & AuthRecipeUserInputOverride;
     linkClickedScreenFeature?: PasswordlessFeatureBaseConfig;
@@ -144,9 +141,9 @@ export type NormalisedConfig = {
 
     override: {
         functions: (
-            originalImplementation: TPPWlessRecipeInterface,
-            builder?: OverrideableBuilder<TPPWlessRecipeInterface>
-        ) => TPPWlessRecipeInterface;
+            originalImplementation: RecipeInterface,
+            builder?: OverrideableBuilder<RecipeInterface>
+        ) => RecipeInterface;
         components: ComponentOverrideMap;
     };
 } & NormalisedAuthRecipeModuleConfig<GetRedirectionURLContext, PreAndPostAPIHookAction, OnHandleEventContext>;
@@ -203,197 +200,3 @@ export type ThirdPartyPasswordlessSignInAndUpThemePropsWithActiveScreen = {
           pwlessChildProps: PwlessChildProps;
       }
 );
-
-export type TPPWlessRecipeInterface = {
-    createCode: (
-        input:
-            | { email: string; userContext: any; options?: RecipeFunctionOptions }
-            | { phoneNumber: string; userContext: any; options?: RecipeFunctionOptions }
-    ) => Promise<{
-        status: "OK";
-        deviceId: string;
-        preAuthSessionId: string;
-        flowType: PasswordlessFlowType;
-        fetchResponse: Response;
-    }>;
-
-    resendCode: (input: {
-        userContext: any;
-        deviceId: string;
-        preAuthSessionId: string;
-        options?: RecipeFunctionOptions;
-    }) => Promise<{
-        status: "OK" | "RESTART_FLOW_ERROR";
-        fetchResponse: Response;
-    }>;
-
-    consumeCode: (
-        input:
-            | {
-                  userInputCode: string;
-                  deviceId: string;
-                  preAuthSessionId: string;
-                  userContext: any;
-                  options?: RecipeFunctionOptions;
-              }
-            | {
-                  preAuthSessionId: string;
-                  linkCode: string;
-                  userContext: any;
-                  options?: RecipeFunctionOptions;
-              }
-    ) => Promise<
-        | {
-              status: "OK";
-              createdUser: boolean;
-              user: PasswordlessUser;
-              fetchResponse: Response;
-          }
-        | {
-              status: "INCORRECT_USER_INPUT_CODE_ERROR" | "EXPIRED_USER_INPUT_CODE_ERROR";
-              failedCodeInputAttemptCount: number;
-              maximumCodeInputAttempts: number;
-              fetchResponse: Response;
-          }
-        | { status: "RESTART_FLOW_ERROR"; fetchResponse: Response }
-    >;
-
-    getPasswordlessLinkCodeFromURL: (input: { userContext: any }) => string;
-
-    getPasswordlessPreAuthSessionIdFromURL: (input: { userContext: any }) => string;
-
-    doesPasswordlessUserEmailExist: (input: {
-        email: string;
-        userContext: any;
-        options?: RecipeFunctionOptions;
-    }) => Promise<{
-        status: "OK";
-        doesExist: boolean;
-        fetchResponse: Response;
-    }>;
-
-    doesPasswordlessUserPhoneNumberExist: (input: {
-        phoneNumber: string;
-        userContext: any;
-        options?: RecipeFunctionOptions;
-    }) => Promise<{
-        status: "OK";
-        doesExist: boolean;
-        fetchResponse: Response;
-    }>;
-
-    // getPasswordlessLoginAttemptInfo: () =>
-    //     | Promise<
-    //           | undefined
-    //           | {
-    //                 deviceId: string;
-    //                 preAuthSessionId: string;
-    //                 contactInfo: string;
-    //                 contactMethod: "EMAIL" | "PHONE";
-    //                 flowType: "USER_INPUT_CODE" | "MAGIC_LINK" | "USER_INPUT_CODE_AND_MAGIC_LINK";
-    //                 lastResend: number;
-    //                 redirectToPath?: string;
-    //             }
-    //       >
-    //     | {
-    //           deviceId: string;
-    //           preAuthSessionId: string;
-    //           contactInfo: string;
-    //           contactMethod: "EMAIL" | "PHONE";
-    //           flowType: "USER_INPUT_CODE" | "MAGIC_LINK" | "USER_INPUT_CODE_AND_MAGIC_LINK";
-    //           lastResend: number;
-    //           redirectToPath?: string;
-    //       }
-    //     | undefined;
-
-    getPasswordlessLoginAttemptInfo: <CustomLoginAttemptInfoProperties>(input: { userContext: any }) =>
-        | Promise<
-              | undefined
-              | ({
-                    deviceId: string;
-                    preAuthSessionId: string;
-                    flowType: PasswordlessFlowType;
-                } & CustomLoginAttemptInfoProperties)
-          >
-        | ({
-              deviceId: string;
-              preAuthSessionId: string;
-              flowType: PasswordlessFlowType;
-          } & CustomLoginAttemptInfoProperties)
-        | undefined;
-
-    // setPasswordlessLoginAttemptInfo: (input: {
-    //     deviceId: string;
-    //     preAuthSessionId: string;
-    //     contactInfo: string;
-    //     contactMethod: "EMAIL" | "PHONE";
-    //     flowType: "USER_INPUT_CODE" | "MAGIC_LINK" | "USER_INPUT_CODE_AND_MAGIC_LINK";
-    //     lastResend: number;
-    //     redirectToPath?: string;
-    // }) => Promise<void> | void;
-
-    setPasswordlessLoginAttemptInfo: <CustomStateProperties>(input: {
-        attemptInfo: {
-            deviceId: string;
-            preAuthSessionId: string;
-            flowType: PasswordlessFlowType;
-        } & CustomStateProperties;
-        userContext: any;
-    }) => Promise<void> | void;
-
-    clearPasswordlessLoginAttemptInfo: (input: { userContext: any }) => Promise<void> | void;
-
-    getAuthorisationURLFromBackend: (input: {
-        providerId: string;
-        userContext: any;
-        options?: RecipeFunctionOptions;
-    }) => Promise<{
-        status: "OK";
-        url: string;
-        fetchResponse: Response;
-    }>;
-
-    thirdPartySignInAndUp: (input: { userContext: any; options?: RecipeFunctionOptions }) => Promise<
-        | {
-              status: "OK";
-              user: UserType;
-              createdNewUser: boolean;
-              fetchResponse: Response;
-          }
-        | {
-              status: "NO_EMAIL_GIVEN_BY_PROVIDER";
-              fetchResponse: Response;
-          }
-    >;
-
-    getStateAndOtherInfoFromStorage: <CustomStateProperties>(input: {
-        userContext: any;
-    }) => (StateObject & CustomStateProperties) | undefined;
-
-    setStateAndOtherInfoToStorage: <CustomStateProperties>(input: {
-        state: StateObject & CustomStateProperties;
-        userContext: any;
-    }) => void;
-
-    getAuthorizationURLWithQueryParamsAndSetState: (input: {
-        providerId: string;
-        authorisationURL: string;
-        userContext: any;
-        providerClientId?: string;
-        options?: RecipeFunctionOptions;
-    }) => Promise<string>;
-
-    generateStateToSendToOAuthProvider: (input: { userContext: any }) => string;
-
-    verifyAndGetStateOrThrowError: <CustomStateProperties>(input: {
-        stateFromAuthProvider: string | undefined;
-        stateObjectFromStorage: (StateObject & CustomStateProperties) | undefined;
-        userContext: any;
-    }) => Promise<StateObject & CustomStateProperties>;
-
-    getAuthCodeFromURL: (input: { userContext: any }) => string;
-
-    getAuthErrorFromURL: (input: { userContext: any }) => string | undefined;
-
-    getAuthStateFromURL: (input: { userContext: any }) => string;
-};